- Fixing warning (signed vs. unsigned comparion). Should find a nicer way to handle this (e.g. use unsigned for sizes always?)

+++ dataLinker/PairedReadConverter.cpp
@@ -99,7 +99,7 @@ PairedReadConverter::PairedReadConverterResult PairedReadConverter::createLinksF
         if (ReadLocations.empty())
             ReadLocations.assign(referenceSize, vector<int>());
         
-        if (ReadLocations.size() != referenceSize)
+        if ((int)ReadLocations.size() != referenceSize)
             result = InconsistentReferenceSets;
         
         if (result == Success)
